+++ app/models/master_person.rb
@@ -5,11 +5,12 @@ class MasterPerson < ApplicationRecord
   has_many :donor_accounts, through: :master_person_donor_accounts
 
   def self.find_or_create_for_person(person, extra = {})
-    if master_person = find_for_person(person, extra = {})
-      return master_person
-    end
+    master_person = find_for_person(person, extra)
+    return master_person if master_person
+
     mp = create!
-    if donor_account = extra[:donor_account]
+    donor_account = extra[:donor_account]
+    if donor_account
       mp.donor_accounts << donor_account
       mp.master_person_sources.create(organization_id: donor_account.organization_id, remote_id: extra[:remote_id]) if extra[:remote_id]
     end
@@ -19,15 +20,14 @@ class MasterPerson < ApplicationRecord
   def self.find_for_person(person, extra = {})
     # Start by looking for a person with the same email address (since that's our one true unique field)
     person.email_addresses.each do |email_address|
-      if other_person = Person.where('people.first_name' => person.first_name,
-                                     'people.last_name' => person.last_name,
-                                     'people.suffix' => person.suffix,
-                                     'email_addresses.email' => email_address.email)
-                              .where('people.middle_name = ? OR people.middle_name is null', person.middle_name)
-                              .joins(:email_addresses)
-                              .first
-        return other_person.master_person
-      end
+      other_person = Person.where('people.first_name' => person.first_name,
+                                  'people.last_name' => person.last_name,
+                                  'people.suffix' => person.suffix,
+                                  'email_addresses.email' => email_address.email)
+                           .where('people.middle_name = ? OR people.middle_name is null', person.middle_name)
+                           .joins(:email_addresses)
+                           .first
+      return other_person.master_person if other_person
     end
 
     # if we have an exact match on name and phone number, that's also pretty good
@@ -47,13 +47,11 @@ class MasterPerson < ApplicationRecord
 
     # If we have a donor account to look at, donor account + name is pretty safe
     if extra[:donor_account]
-      if other_person = extra[:donor_account].people.where('people.first_name' => person.first_name,
-                                                           'people.last_name' => person.last_name,
-                                                           'people.suffix' => person.suffix)
-                                             .find_by('people.middle_name = ? OR people.middle_name is null', person.middle_name)
-        return other_person.master_person
-      end
-
+      other_person = extra[:donor_account].people.where('people.first_name' => person.first_name,
+                                                        'people.last_name' => person.last_name,
+                                                        'people.suffix' => person.suffix)
+                                          .find_by('people.middle_name = ? OR people.middle_name is null', person.middle_name)
+      return other_person.master_person if other_person
     end
 
     nil