+++ app/models/prayer_letters_account.rb
@@ -40,8 +40,9 @@ class PrayerLettersAccount < ApplicationRecord
     contacts = JSON.parse(get_response(:get, '/api/v1/contacts'))['contacts']
 
     contacts.each do |pl_contact|
-      next unless pl_contact['external_id'] &&
-                  contact = account_list.contacts.find_by(id: pl_contact['external_id'])
+      next unless pl_contact['external_id']
+      contact = account_list.contacts.find_by(id: pl_contact['external_id'])
+      next unless contact
 
       contact.update_columns(prayer_letters_id: pl_contact['contact_id'],
                              prayer_letters_params: contact_params_map ? contact_params_map[contact.id] : nil)